v1.0
This is the version as it sits in neocities in June 2019.

It's the same as the version with which I did some user testing (c 50 users via google ads)

In this version I've added in mixpanel code snippets in index.html and the tio js file.

I will need to add in the netlify code snippets (not done yet)