Added name attribute to form tags
When I tried implementing the previous commit it didn't work, but it said here (https://cobwwweb.com/what-you-need-to-know-about-netlify-forms) that I need a "name" attribute in the form tags (even though it didn't say this in netlify's own website: https://app.netlify.com/sites/talk-it-over/forms). So I'm trying adding in the name attribute in both forms. No other changes